+++ StevesScriptorium.psm1
@@ -62,6 +62,13 @@ function global:toolkit {
 
         # Groups
         "get-groupmembers"       = "List all members of a group"
+
+        # MFA & Auth
+        "get-smsmfa"             = "Show SMS/phone MFA methods for a user"
+        "set-smsmfa"             = "Update the phone number on an existing SMS MFA method"
+        "add-smsmfa"             = "Register a new SMS/phone MFA method for a user"
+        "add-tap"                = "Create a Temporary Access Pass for a user"
+        "remove-taps"            = "Remove all Temporary Access Passes for a user"
     }
 
     $sectionHeaders = @{
@@ -70,6 +77,7 @@ function global:toolkit {
         "get-tenantreport"       = "Tenant Health"
         "check-mailflow"         = "Mailbox & Exchange"
         "get-groupmembers"       = "Groups"
+        "get-smsmfa"             = "MFA & Auth"
     }
 
     if (-not $Command) {